# api-docs.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"baguette-item-controller","description":"Baguette Item Controller"},{"name":"baguette-order-controller","description":"Baguette Order Controller"},{"name":"basic-error-controller","description":"Basic Error Controller"},{"name":"customer-controller","description":"Customer Controller"},{"name":"home-controller","description":"Home Controller"},{"name":"ingredient-controller","description":"Ingredient Controller"},{"name":"ingredient-type-controller","description":"Ingredient Type Controller"},{"name":"item-controller","description":"Item Controller"},{"name":"operation-handler","description":"Operation Handler"},{"name":"special-offer-controller","description":"Special Offer Controller"},{"name":"web-mvc-links-handler","description":"Web Mvc Links Handler"}],"paths":{"/":{"get":{"tags":["home-controller"],"summary":"home","operationId":"homeUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator":{"get":{"tags":["web-mvc-links-handler"],"summary":"links","operationId":"linksUsingGET","produces":["application/json","application/vnd.spring-boot.actuator.v2+json","application/vnd.spring-boot.actuator.v3+json"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object","additionalProperties":{"$ref":"#/definitions/Link"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/mappings":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET","produces":["application/json","application/vnd.spring-boot.actuator.v2+json","application/vnd.spring-boot.actuator.v3+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/afterLogin":{"get":{"tags":["home-controller"],"summary":"afterLogin","operationId":"afterLoginUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/baguetteItem/all/{baguetteOrderId}":{"get":{"tags":["baguette-item-controller"],"summary":"getBaguetteItems","operationId":"getBaguetteItemsUsingGET","produces":["*/*"],"parameters":[{"name":"baguetteOrderId","in":"path","description":"baguetteOrderId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/BaguetteItem"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/baguetteItem/create/{baguetteOrderId}":{"post":{"tags":["baguette-item-controller"],"summary":"createBaguetteItem","operationId":"createBaguetteItemUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"baguetteOrderId","in":"path","description":"baguetteOrderId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/BaguetteItem"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/baguetteItem/create/{baguetteOrderId}/{specialOfferId}":{"post":{"tags":["baguette-item-controller"],"summary":"createBaguetteItemFromSpecialOffer","operationId":"createBaguetteItemFromSpecialOfferUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"baguetteOrderId","in":"path","description":"baguetteOrderId","required":true,"type":"integer","format":"int64"},{"name":"specialOfferId","in":"path","description":"specialOfferId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/BaguetteItem"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/baguetteItem/detail/{baguetteItemId}":{"get":{"tags":["baguette-item-controller"],"summary":"getBaguetteItem","operationId":"getBaguetteItemUsingGET","produces":["*/*"],"parameters":[{"name":"baguetteItemId","in":"path","description":"baguetteItemId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/BaguetteItem"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/baguetteItem/remove/{baguetteItemId}":{"delete":{"tags":["baguette-item-controller"],"summary":"removeBaguetteItem","operationId":"removeBaguetteItemUsingDELETE","produces":["*/*"],"parameters":[{"name":"baguetteItemId","in":"path","description":"baguetteItemId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/baguetteOrder/actual":{"get":{"tags":["baguette-order-controller"],"summary":"getBaguetteOrderActual","operationId":"getBaguetteOrderActualUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/BaguetteOrder"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/baguetteOrder/all":{"get":{"tags":["baguette-order-controller"],"summary":"getBaguetteOrders","operationId":"getBaguetteOrdersUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/BaguetteOrder"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/baguetteOrder/all/{state}":{"get":{"tags":["baguette-order-controller"],"summary":"getBaguetteOrdersByState","operationId":"getBaguetteOrdersByStateUsingGET","produces":["*/*"],"parameters":[{"name":"state","in":"path","description":"state","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/BaguetteOrder"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/baguetteOrder/confirm/{baguetteOrderId}":{"put":{"tags":["baguette-order-controller"],"summary":"confirmBaguetteOrder","operationId":"confirmBaguetteOrderUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"baguetteOrderId","in":"path","description":"baguetteOrderId","required":true,"type":"integer","format":"int64"},{"name":"date","in":"query","description":"date","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/baguetteOrder/create":{"post":{"tags":["baguette-order-controller"],"summary":"createBaguetteOrder","operationId":"createBaguetteOrderUsingPOST","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/BaguetteOrder"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/baguetteOrder/detail/{baguetteOrderId}":{"get":{"tags":["baguette-order-controller"],"summary":"getBaguetteOrder","operationId":"getBaguetteOrderUsingGET","produces":["*/*"],"parameters":[{"name":"baguetteOrderId","in":"path","description":"baguetteOrderId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/BaguetteOrder"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/baguetteOrder/remove/{baguetteOrderId}":{"delete":{"tags":["baguette-order-controller"],"summary":"removeBaguetteOrder","operationId":"removeBaguetteOrderUsingDELETE","produces":["*/*"],"parameters":[{"name":"baguetteOrderId","in":"path","description":"baguetteOrderId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/baguetteOrder/update/{baguetteOrderId}":{"put":{"tags":["baguette-order-controller"],"summary":"updateBaguetteOrder","operationId":"updateBaguetteOrderUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"baguetteOrderId","in":"path","description":"baguetteOrderId","required":true,"type":"integer","format":"int64"},{"name":"note","in":"query","description":"note","required":false,"type":"string"}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/customer/all":{"get":{"tags":["customer-controller"],"summary":"getCustomers","operationId":"getCustomersUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Customer"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/customer/create":{"post":{"tags":["customer-controller"],"summary":"registerNewCustomer","operationId":"registerNewCustomerUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"customer","description":"customer","required":true,"schema":{"$ref":"#/definitions/Customer"}}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/customer/update/{customerId}":{"put":{"tags":["customer-controller"],"summary":"updateCustomer","operationId":"updateCustomerUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"customerId","in":"path","description":"customerId","required":true,"type":"integer","format":"int64"},{"name":"email","in":"query","description":"email","required":false,"type":"string"},{"name":"emailShort","in":"query","description":"emailShort","required":false,"type":"string"},{"name":"name","in":"query","description":"name","required":false,"type":"string"}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/detail":{"get":{"tags":["home-controller"],"summary":"detail","operationId":"detailUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/error":{"get":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"head":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingHEAD","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"post":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPOST","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPUT","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingDELETE","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"options":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingOPTIONS","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"patch":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPATCH","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/ingredient/all":{"get":{"tags":["ingredient-controller"],"summary":"getIngredients","operationId":"getIngredientsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Ingredient"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/ingredient/detail/{ingredientName}":{"get":{"tags":["ingredient-controller"],"summary":"getIngredientByName","operationId":"getIngredientByNameUsingGET","produces":["*/*"],"parameters":[{"name":"ingredientName","in":"path","description":"ingredientName","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Ingredient"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/ingredientType/all":{"get":{"tags":["ingredient-type-controller"],"summary":"getIngredientTypes","operationId":"getIngredientTypesUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/IngredientType"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/item/create/{baguetteItemId}":{"post":{"tags":["item-controller"],"summary":"createNewItem","operationId":"createNewItemUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"baguetteItemId","in":"path","description":"baguetteItemId","required":true,"type":"integer","format":"int64"},{"in":"body","name":"item","description":"item","required":true,"schema":{"$ref":"#/definitions/Item"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Item"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/item/remove/{itemId}":{"delete":{"tags":["item-controller"],"summary":"removeItem","operationId":"removeItemUsingDELETE","produces":["*/*"],"parameters":[{"name":"itemId","in":"path","description":"itemId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/item/update/{itemId}":{"put":{"tags":["item-controller"],"summary":"updateItem","operationId":"updateItemUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"amount","in":"query","description":"amount","required":false,"type":"integer","format":"int32"},{"name":"itemId","in":"path","description":"itemId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/item/{baguetteItemId}/all":{"get":{"tags":["item-controller"],"summary":"getItems","operationId":"getItemsUsingGET","produces":["*/*"],"parameters":[{"name":"baguetteItemId","in":"path","description":"baguetteItemId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Item"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/specialOffer/all":{"get":{"tags":["special-offer-controller"],"summary":"getSpecialOffers","operationId":"getSpecialOffersUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/SpecialOffer"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/specialOffer/all/{active}":{"get":{"tags":["special-offer-controller"],"summary":"getSpecialOffersByActive","operationId":"getSpecialOffersByActiveUsingGET","produces":["*/*"],"parameters":[{"name":"active","in":"path","description":"active","required":true,"type":"boolean"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/SpecialOffer"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/specialOffer/detail/{specialOfferId}":{"get":{"tags":["special-offer-controller"],"summary":"getSpecialOffer","operationId":"getSpecialOfferUsingGET","produces":["*/*"],"parameters":[{"name":"specialOfferId","in":"path","description":"specialOfferId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/SpecialOffer"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"BaguetteItem":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"items":{"type":"array","items":{"$ref":"#/definitions/Item"}},"offer":{"type":"boolean"},"price":{"type":"number","format":"double"}},"title":"BaguetteItem"},"BaguetteOrder":{"type":"object","properties":{"baguetteItems":{"type":"array","items":{"$ref":"#/definitions/BaguetteItem"}},"customer":{"$ref":"#/definitions/Customer"},"date":{"type":"string","example":"dd-MM-yyyy, HH:mm:ss"},"id":{"type":"integer","format":"int64"},"note":{"type":"string"},"price":{"type":"number","format":"double"},"state":{"type":"integer","format":"int32"}},"title":"BaguetteOrder"},"Customer":{"type":"object","properties":{"email":{"type":"string"},"emailShort":{"type":"string"},"id":{"type":"integer","format":"int64"},"name":{"type":"string"}},"title":"Customer"},"Ingredient":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"ingredientType":{"$ref":"#/definitions/IngredientType"},"name":{"type":"string"},"price":{"type":"number","format":"double"}},"title":"Ingredient"},"IngredientType":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}},"title":"IngredientType"},"Item":{"type":"object","properties":{"amount":{"type":"integer","format":"int32"},"id":{"type":"integer","format":"int64"},"ingredient":{"$ref":"#/definitions/Ingredient"},"name":{"type":"string"},"price":{"type":"number","format":"double"}},"title":"Item"},"Link":{"type":"object","properties":{"href":{"type":"string"},"templated":{"type":"boolean"}},"title":"Link"},"ModelAndView":{"type":"object","properties":{"empty":{"type":"boolean"},"model":{"type":"object"},"modelMap":{"type":"object","additionalProperties":{"type":"object"}},"reference":{"type":"boolean"},"status":{"type":"string","enum":["ACCEPTED","ALREADY_REPORTED","BAD_GATEWAY","BAD_REQUEST","BANDWIDTH_LIMIT_EXCEEDED","CHECKPOINT","CONFLICT","CONTINUE","CREATED","DESTINATION_LOCKED","EXPECTATION_FAILED","FAILED_DEPENDENCY","FORBIDDEN","FOUND","GATEWAY_TIMEOUT","GONE","HTTP_VERSION_NOT_SUPPORTED","IM_USED","INSUFFICIENT_SPACE_ON_RESOURCE","INSUFFICIENT_STORAGE","INTERNAL_SERVER_ERROR","I_AM_A_TEAPOT","LENGTH_REQUIRED","LOCKED","LOOP_DETECTED","METHOD_FAILURE","METHOD_NOT_ALLOWED","MOVED_PERMANENTLY","MOVED_TEMPORARILY","MULTIPLE_CHOICES","MULTI_STATUS","NETWORK_AUTHENTICATION_REQUIRED","NON_AUTHORITATIVE_INFORMATION","NOT_ACCEPTABLE","NOT_EXTENDED","NOT_FOUND","NOT_IMPLEMENTED","NOT_MODIFIED","NO_CONTENT","OK","PARTIAL_CONTENT","PAYLOAD_TOO_LARGE","PAYMENT_REQUIRED","PERMANENT_REDIRECT","PRECONDITION_FAILED","PRECONDITION_REQUIRED","PROCESSING","PROXY_AUTHENTICATION_REQUIRED","REQUESTED_RANGE_NOT_SATISFIABLE","REQUEST_ENTITY_TOO_LARGE","REQUEST_HEADER_FIELDS_TOO_LARGE","REQUEST_TIMEOUT","REQUEST_URI_TOO_LONG","RESET_CONTENT","SEE_OTHER","SERVICE_UNAVAILABLE","SWITCHING_PROTOCOLS","TEMPORARY_REDIRECT","TOO_EARLY","TOO_MANY_REQUESTS","UNAUTHORIZED","UNAVAILABLE_FOR_LEGAL_REASONS","UNPROCESSABLE_ENTITY","UNSUPPORTED_MEDIA_TYPE","UPGRADE_REQUIRED","URI_TOO_LONG","USE_PROXY","VARIANT_ALSO_NEGOTIATES"]},"view":{"$ref":"#/definitions/View"},"viewName":{"type":"string"}},"title":"ModelAndView"},"SpecialOffer":{"type":"object","properties":{"active":{"type":"boolean"},"id":{"type":"integer","format":"int64"},"items":{"type":"array","items":{"$ref":"#/definitions/Item"}},"name":{"type":"string"},"price":{"type":"number","format":"double"}},"title":"SpecialOffer"},"View":{"type":"object","properties":{"contentType":{"type":"string"}},"title":"View"}}}